artifact: fix failing tests for sandbox inspection
In #27398 we changed the way we used `os.Root` to protect the client when
merging artifact directories to fix failing tests on Windows where this approach
would always fail with "invalid argument". But this has now started failing on
unit tests in CI, but only on Linux (where we'll usually have Landlock anyways),
specifically on AMD64. While we try to investigate, let's revert #27398 but only
on non-Windows.

Ref: https://github.com/hashicorp/nomad/pull/27398
