+++ client/allocrunner/taskrunner/getter/sandbox_test.go
@@ -12,7 +12,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"syscall"
 	"testing"
 	"time"
 
@@ -89,32 +88,6 @@ func TestSandbox_Get_insecure_http(t *testing.T) {
 	artifact.GetterInsecure = true
 	err = sbox.Get(env, artifact, "nobody")
 	must.NoError(t, err)
-}
-
-func TestSandbox_Get_chown(t *testing.T) {
-	testutil.RequireRoot(t) // NOTE: required for chown call
-	logger := testlog.HCLogger(t)
-
-	ac := artifactConfig(10 * time.Second)
-	sbox := New(ac, logger)
-
-	_, taskDir := SetupDir(t)
-	env := noopTaskEnv(taskDir)
-
-	artifact := &structs.TaskArtifact{
-		GetterSource: "https://raw.githubusercontent.com/hashicorp/go-set/main/go.mod",
-		RelativeDest: "local/downloads",
-		Chown:        true,
-	}
-
-	err := sbox.Get(env, artifact, "nobody")
-	must.NoError(t, err)
-
-	info, err := os.Stat(filepath.Join(taskDir, "local", "downloads"))
-	must.NoError(t, err)
-
-	uid := info.Sys().(*syscall.Stat_t).Uid
-	must.Eq(t, 65534, uid) // nobody's conventional uid
 }
 
 func TestSandbox_Get_inspection(t *testing.T) {
@@ -219,32 +192,6 @@ func TestSandbox_Get_inspection(t *testing.T) {
 
 		_, err = os.Stat(filepath.Join(taskDir, "local", "downloads", "test-file"))
 		must.NoError(t, err)
-	})
-
-	t.Run("properly chowns destination", func(t *testing.T) {
-		taskDir, sbox, env := sandboxSetup()
-		src, _ := servTestFile(t, "test-file")
-
-		artifact := &structs.TaskArtifact{
-			GetterSource: src,
-			RelativeDest: "local/downloads",
-			Chown:        true,
-		}
-
-		err := sbox.Get(env, artifact, "nobody")
-		must.NoError(t, err)
-
-		info, err := os.Stat(filepath.Join(taskDir, "local", "downloads"))
-		must.NoError(t, err)
-
-		uid := info.Sys().(*syscall.Stat_t).Uid
-		must.Eq(t, 65534, uid) // nobody's conventional uid
-
-		info, err = os.Stat(filepath.Join(taskDir, "local", "downloads", "test-file"))
-		must.NoError(t, err)
-
-		uid = info.Sys().(*syscall.Stat_t).Uid
-		must.Eq(t, 65534, uid) // nobody's conventional uid
 	})
 
 	t.Run("when destination file exists", func(t *testing.T) {
